Permutation Group implementation (stand-alone code) + Travelling salesman example
A stand-alone implementation of the algorithm on the Permutation Lattice has been implemented!
The idea of information propagation is reproposed on the graph induced by the permutations (where the edges denote a 1-swap distance between two permutations).
This very basic experiment shows how the application of the algorithm to the procedure of optimization can improve the solution.